+++ FuturaeKit.framework/Headers/FTRNotificationDelegate.h
@@ -13,8 +13,7 @@
 #import <Foundation/Foundation.h>
 
 /**
- * The `FTRNotificationDelegate`'s methods allow the delegate to be informed when SoundProof and
- * approve authentication attempts or user unenrolled actions are received via Push Notifications.
+ * The `FTRNotificationDelegate`'s methods allow the delegate to be informed when approve authentication attempts or user unenrolled actions are received via Push Notifications.
  */
 @protocol FTRNotificationDelegate <NSObject>